Prevent adding repeat users to a group.
